fixed bugs in seeded data causing script to throw errors
